training-platform core: extract training codegen helpers to codeGenerateTraining.py
testUtils/codeGenerate.py had accumulated ~880 lines of training / optimizer
codegen helpers piled on top of the upstream inference module, plus a dead
"Initialize all output buffers to zero" block that had crept into
generateTestNetworkImplementation (it declared local variables and looped
over graph outputs without ever emitting anything into retStr).  The net
effect was that this PR's diff in codeGenerate.py was +883 lines and an
interleaving that made the inference vs training boundary invisible to
reviewers.

Reset testUtils/codeGenerate.py to byte-identical upstream/devel and move
every training-side helper to a new testUtils/codeGenerateTraining.py:

  - generateTrainingTestInputsHeader / generateTrainingTestOutputsHeader
  - generateTrainingNetworkHeader / generateTrainingNetworkImplementation
  - generateTrainingTestNetwork
  - build_shared_buffer_maps / _patch_shared_buffers / _patch_shared_arenas
  - _ensure_training_l1_capacity
  - generateOptimizerNetworkHeader / generateOptimizerNetworkImplementation
  - generateOptimizerTestNetwork

codeGenerateTraining.py imports generateL3HexDump from codeGenerate (the
only inference helper the training functions reuse), which is a clean
forward dependency.  No back-edge from codeGenerate to training.

The four codegen entry points are updated to pull their training /
optimizer helpers from testUtils.codeGenerateTraining instead of
testUtils.codeGenerate.  No other files change.

Verified on Siracusa: simplemlp_train passes 0/4 (diff=0.000000 at every
step) in both non-tiled and tiled runs.
