+++ DeeployTest/generateOptimizerNetwork.py
@@ -27,7 +27,7 @@ from pathlib import Path
 
 import onnx
 import onnx_graphsurgeon as gs
-from testUtils.codeGenerate import build_shared_buffer_maps, generateOptimizerTestNetwork
+from testUtils.codeGenerateTraining import build_shared_buffer_maps, generateOptimizerTestNetwork
 from testUtils.platformMapping import mapDeployer, mapPlatform, setupMemoryPlatform
 from testUtils.testRunner import TestGeneratorArgumentParser
 from testUtils.trainingUtils import add_cores_arg, add_memory_level_args, add_optimizer_training_dir_arg, \
